fixed consulattions

+++ nwaHospitalApi/src/main/java/com/nwaHospitalApi/nwaHospitalApi/Repositories/SurgeryRepository.java
@@ -4,6 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.nwaHospitalApi.nwaHospitalApi.Entities.surgery;
 
-public interface SurgeryRepository extends JpaRepository<surgery, Long> {
+public interface SurgeryRepository extends JpaRepository<surgery, Integer> {
+
 }
